requirements.txt
Django==5.2.5
google-generativeai>=0.5.0
python-dotenv>=1.0.1
